adds documentation and simplifies targets
this was returning an array before in part because it needed to check if the array was empty to detect an error case.  now, it doesn't need to do that, so there's no point in returning an array that statically only has one item just to join it with commas later (which will have no effect since it only has one item).  If we ever want to modify this, we can just as well change something like `'node12-macos-x64'` to `'node12-macos-x64,node12-macos-arm64'` in the `getTargets` function.